backend/core/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.5
django-js-asset==3.1.2
django-mptt==0.17.0
django-rest-swagger==2.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
httplib2==0.22.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauth2client==4.1.3
openapi-codec==1.3.2
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
